isolated dcm code path so that overall refactoring is not blocked on it
